Add SQLite database and message history function
This commit implements a SQLite database and initializes it to store chat messages. It also adds a feature to load and display all previous messages to users upon their connection to the chat. This change aims to enhance user experience by creating a history of messages.

TODO : Fix the double send probleme and the reconection problem also lets add usernames and better visuals